requirements.txt
asgiref==3.2.7
Django==3.0.6
django-crispy-forms==1.9.1
gunicorn==20.0.4
Pillow==7.1.2
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
